Update __init__.py (#3)
Change made to handle case where an utterance contains multiple sentences.  Mimic1 resets the duration back to zero when a new sentence is started.  This fix detects that and then resets the start timer to the current time.  Otherwise, the time.time()<start+end test would fail and the routine would end early.